refactor : pages와 templates 분리하기 [#34]
